Copy-Move Forgery
Copy-Move Forgery in Computer Forensics.
Summary and introduction to the problem to be solved: I want to detect Tampered images from the given database. Detection of copy-move forgery in digital images. Copy-move forgery is a common form of image manipulation where a part of an image is copied and pasted within the same image to create a duplicate or altered region.